Add Radio Not Found Exception and rearrange test files
