+++ src/main/java/com/mgmtp/radio/controller/RadioResponseExceptionHandler.java
@@ -2,8 +2,10 @@ package com.mgmtp.radio.controller;
 
 import com.mgmtp.radio.controller.response.RadioErrorResponse;
 import com.mgmtp.radio.exception.RadioBadRequestException;
+import com.mgmtp.radio.exception.RadioNotFoundException;
 import com.mgmtp.radio.exception.RadioServiceException;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.core.env.Environment;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,9 +19,15 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExcep
 @ControllerAdvice
 public class RadioResponseExceptionHandler extends ResponseEntityExceptionHandler {
 
+    private final Environment env;
+
+    public RadioResponseExceptionHandler(Environment env) {
+        this.env = env;
+    }
+
     // return 500 INTERNAL_SERVER_ERROR
     @ExceptionHandler({Exception.class})
-    public ResponseEntity<Object> handleNotFoundException(Exception exception, WebRequest webRequest){
+    public ResponseEntity<Object> handleInternalServerErrorException(Exception exception, WebRequest webRequest){
         log.error("Exception occurred when processing request", exception);
         return new ResponseEntity<>("Resource Not Found", new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
     }
@@ -45,5 +53,20 @@ public class RadioResponseExceptionHandler extends ResponseEntityExceptionHandle
         }
 
         return new ResponseEntity<>(new RadioErrorResponse(message), new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    // return 404 NOT_FOUND
+    @ExceptionHandler(RadioNotFoundException.class)
+    protected ResponseEntity<Object> handleNotFoundException(RadioNotFoundException exception, WebRequest webRequest) {
+        log.error("Exception processing request", exception);
+
+        String message;
+
+        if (StringUtils.isEmpty(exception.getMessage())) {
+            message = env.getProperty("exception.not_found");
+        } else {
+            message = exception.getMessage();
+        }
+        return new ResponseEntity<>(new RadioErrorResponse(message), new HttpHeaders(), HttpStatus.NOT_FOUND);
     }
 }
