+++ src/main/java/com/mgmtp/radio/service/user/FavoriteSongServiceImpl.java
@@ -41,7 +41,7 @@ public class FavoriteSongServiceImpl implements FavoriteSongService {
 	}
 
 	@Override
-	public Mono<FavoriteSongDTO> delete(String favoriteSongId, String userId) throws RadioNotFoundException {
+	public Mono<FavoriteSongDTO> delete(String favoriteSongId, String userId) {
 		return favoriteSongRepository.findByIdAndUserId(favoriteSongId, userId).flatMap(song -> {
 			return favoriteSongRepository.delete(song).then(Mono.just(favoriteSongMapper.favoriteSongToFavoriteSongDTO(song)));
 		}).switchIfEmpty(Mono.error(new RadioNotFoundException()));
