+++ src/main/resources/message.properties
@@ -1,3 +1,5 @@
-user.favorite.error.exist.userId= "User id does not exist"
-user.favorite.error.exist.songId= "Song id does not exist"
-user.favorite.error.exist.song= "Favorite song has already existed"
+user.favorite.error.exist.userId=User id does not exist
+user.favorite.error.exist.songId=Song id does not exist
+user.favorite.error.exist.song=Favorite song has already existed
+
+exception.not_found=Radio resource could not be found but may be available in the future.
