+++ src/test/java/com/mgmtp/radio/service/user/favorite/DeleteFavoriteSongImplTest.java
@@ -1,7 +1,8 @@
-package com.mgmtp.radio.service.user.service.favorite;
+package com.mgmtp.radio.service.user.favorite;
 
 import com.mgmtp.radio.domain.user.FavoriteSong;
 import com.mgmtp.radio.dto.user.FavoriteSongDTO;
+import com.mgmtp.radio.exception.RadioNotFoundException;
 import com.mgmtp.radio.mapper.user.FavoriteSongMapper;
 import com.mgmtp.radio.respository.user.FavoriteSongRepository;
 import com.mgmtp.radio.service.user.FavoriteSongService;
@@ -10,16 +11,11 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
-import java.util.UUID;
+import static org.mockito.Mockito.*;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.when;
-
-public class GetFavoriteSongByUserIdImplTest {
+public class DeleteFavoriteSongImplTest {
 
     @Mock
     FavoriteSongRepository favoriteSongRepository;
@@ -35,10 +31,10 @@ public class GetFavoriteSongByUserIdImplTest {
     }
 
     @Test
-    public void getFavoriteListByUserIdSuccess() {
+    public void deleteSuccess() {
         //given
         FavoriteSongDTO favoriteSongDTO = new FavoriteSongDTO();
-        favoriteSongDTO.setId(UUID.randomUUID().toString());
+        favoriteSongDTO.setId("001");
         favoriteSongDTO.setUserId("001");
         favoriteSongDTO.setSongId("001");
 
@@ -47,30 +43,33 @@ public class GetFavoriteSongByUserIdImplTest {
         favoriteSong.setUserId(favoriteSongDTO.getUserId());
         favoriteSong.setSongId(favoriteSongDTO.getSongId());
 
-        when(favoriteSongRepository.findByUserId(anyString())).thenReturn(Flux.just(favoriteSong));
+        when(favoriteSongRepository.findByIdAndUserId(anyString(), anyString())).thenReturn(Mono.just(favoriteSong));
 
         // when
-        Flux<FavoriteSongDTO> result = favoriteSongService.findByUserId(favoriteSongDTO.getUserId());
-        FavoriteSongDTO expected = result.log().next().block();
+        favoriteSongRepository.delete(favoriteSong);
 
-        // then
-        assertEquals(favoriteSongDTO, expected);
+        //then
+        verify(favoriteSongRepository, times(1)).delete(any(FavoriteSong.class));
     }
 
-    @Test
-    public void getEmptyFavoriteListByUserId() {
+    @Test(expected = Exception.class)
+    public void deleteFailureWithWrongUserId() {
         //given
         FavoriteSongDTO favoriteSongDTO = new FavoriteSongDTO();
-        favoriteSongDTO.setId(UUID.randomUUID().toString());
+        favoriteSongDTO.setId("001");
         favoriteSongDTO.setUserId("001");
         favoriteSongDTO.setSongId("001");
 
-        when(favoriteSongRepository.findByUserId(anyString())).thenReturn(Flux.empty());
+        FavoriteSong favoriteSong = new FavoriteSong();
+        favoriteSong.setId(favoriteSongDTO.getId());
+        favoriteSong.setUserId(favoriteSongDTO.getUserId());
+        favoriteSong.setSongId(favoriteSongDTO.getSongId());
+
+
+        when(favoriteSongRepository.findByIdAndUserId(anyString(), anyString())).thenThrow(new RadioNotFoundException());
+        when(favoriteSongRepository.delete(any(FavoriteSong.class)));
 
         // when
-        Flux<FavoriteSongDTO> result = favoriteSongService.findByUserId(favoriteSongDTO.getUserId());
-
-        // then
-        assertNull(result.log().next().block());
+        favoriteSongService.delete(favoriteSongDTO.getId(), favoriteSongDTO.getUserId());
     }
 }
